pyproject.toml
[project]
name = "ArSLRecognition"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "opencv-python>=4.10.0.84",
    "pillow>=10.4.0",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "torchvision>=0.26.0",
    "ultralytics>=8.3.0",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu128"
